refactor(backups): split provider API responsibilities
